# apps/admin-web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Duolingo Unofficial API (Main)",
    "description": "Unofficial API description for Duolingo.",
    "version": "1.0.0"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/2017-06-30/users/{userID}": {
      "parameters": [
        {
          "in": "path",
          "name": "userID",
          "required": true,
          "description": "User ID",
          "schema": {
            "type": "integer",
            "minimum": 1
          }
        },
        {
          "in": "query",
          "name": "fields",
          "description": "Fields to retrieve",
          "style": "form",
          "explode": false,
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      ],
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "setLanguage",
        "summary": "Set User Language",
        "description": "Set User Language/Course",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseId": {
                    "type": "string",
                    "example": "DUOLINGO_KO_EN"
                  },
                  "fromLanguage": {
                    "type": "string",
                    "example": "en"
                  },
                  "learningLanguage": {
                    "type": "string",
                    "example": "ko"
                  }
                }
              }
            }
          },
          "description": "The user to create."
        },
        "responses": {
          "200": {
            "description": "An empty object"
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "getUser",
        "summary": "Returns user information.",
        "description": "User information.",
        "responses": {
          "200": {
            "description": "A User object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "The specified user ID is invalid (e.g. not a number)."
          },
          "404": {
            "description": "A user with the specified ID was not found."
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://www.duolingo.com"
    },
    {
      "url": "https://www.duolingo.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "Achievement": {
        "properties": {
          "tier": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          },
          "tierCounts": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "name": {
            "type": "string"
          },
          "shouldShowUnlock": {
            "type": "boolean"
          }
        }
      },
      "OptionalFeature": {
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "SmartTip": {
        "properties": {
          "url": {
            "type": "string"
          },
          "smartTipId": {
            "type": "string"
          }
        }
      },
      "definitions-TrackingProperties": {
        "properties": {
          "direction": {
            "type": "string"
          },
          "total_crowns": {
            "type": "integer"
          },
          "took_placementtest": {
            "type": "boolean"
          },
          "learning_language": {
            "type": "string"
          },
          "gold_skill_count": {
            "type": "integer"
          },
          "gold_tree_percent": {
            "type": "number"
          },
          "num_skills_newly_decayed": {
            "type": "integer"
          },
          "num_skills_decayed": {
            "type": "integer"
          },
          "ui_language": {
            "type": "string"
          },
          "max_tree_level": {
            "type": "integer"
          }
        }
      },
      "ProgressQuizItem": {
        "properties": {
          "endTime": {
            "type": "integer"
          },
          "score": {
            "type": "number"
          },
          "startTime": {
            "type": "integer"
          }
        }
      },
      "CourseSection": {
        "properties": {
          "numRows": {
            "type": "integer"
          },
          "checkpointFinished": {
            "type": "boolean"
          },
          "checkpointAccessible": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "preload": {
            "type": "boolean"
          },
          "extraCrowns": {
            "type": "integer"
          },
          "placementTestAvailable": {
            "type": "boolean"
          },
          "authorId": {
            "type": "string"
          },
          "progressVersion": {
            "type": "integer"
          }
        }
      },
      "CurrentCourse": {
        "properties": {
          "status": {
            "type": "string"
          },
          "learningLanguage": {
            "type": "string"
          },
          "healthEnabled": {
            "type": "boolean"
          },
          "xp": {
            "type": "integer"
          },
          "wordsLearned": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "fromLanguage": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "smartTips": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SmartTip"
            }
          },
          "trackingProperties": {
            "$ref": "#/components/schemas/definitions-TrackingProperties"
          },
          "crowns": {
            "type": "integer"
          },
          "progressQuizHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgressQuizItem"
            }
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseSection"
            }
          }
        }
      },
      "Course": {
        "properties": {
          "id": {
            "type": "string"
          },
          "preload": {
            "type": "boolean"
          },
          "placementTestAvailable": {
            "type": "boolean"
          },
          "authorId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "learningLanguage": {
            "type": "string"
          },
          "xp": {
            "type": "integer"
          },
          "healthEnabled": {
            "type": "boolean"
          },
          "fromLanguage": {
            "type": "string"
          },
          "crowns": {
            "type": "integer"
          }
        }
      },
      "XpGain": {
        "properties": {
          "eventType": {
            "type": "string"
          },
          "xp": {
            "type": "integer"
          },
          "skillId": {
            "type": "string"
          },
          "time": {
            "type": "integer"
          }
        }
      },
      "StreakData": {
        "properties": {
          "length": {
            "type": "integer"
          },
          "updatedTimeZone": {
            "type": "string"
          },
          "xpGoal": {
            "type": "integer"
          },
          "updatedTimestamp": {
            "type": "integer"
          }
        }
      },
      "ReferralInfo": {
        "properties": {
          "inviterName": {
            "type": "string"
          },
          "unconsumedInviteeName": {
            "type": "string"
          },
          "numBonusesReady": {
            "type": "integer"
          },
          "hasReachedCap": {
            "type": "boolean"
          },
          "isEligibleForBonus": {
            "type": "boolean"
          }
        }
      },
      "definitions-LastStreak": {
        "properties": {
          "isAvailableForRepair": {
            "type": "boolean"
          },
          "googlePlayProductId": {
            "type": "string"
          },
          "googlePlayDevPayload": {
            "type": "string"
          },
          "length": {
            "type": "integer"
          },
          "daysAgo": {
            "type": "integer"
          },
          "lastReachedGoal": {
            "type": "integer"
          },
          "shortenedProductId": {
            "type": "string"
          }
        }
      },
      "ShopItem": {
        "properties": {
          "wagerDay": {
            "type": "integer"
          },
          "purchaseId": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "integer"
          },
          "itemName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "purchasePrice": {
            "type": "integer"
          }
        }
      },
      "Health": {
        "properties": {
          "maxHearts": {
            "type": "integer"
          },
          "healthEnabled": {
            "type": "boolean"
          },
          "eligibleForFreeRefill": {
            "type": "boolean"
          },
          "secondsPerHeartSegment": {
            "type": "integer"
          },
          "hearts": {
            "type": "integer"
          },
          "useHealth": {
            "type": "boolean"
          },
          "secondsUntilNextHeartSegment": {
            "type": "integer"
          }
        }
      },
      "User": {
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "monthlyXp": {
            "type": "integer"
          },
          "streak": {
            "type": "integer"
          },
          "motivation": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "learningLanguage": {
            "type": "string"
          },
          "hasFacebookId": {
            "type": "boolean"
          },
          "facebookId": {
            "type": "string"
          },
          "inviteURL": {
            "type": "string"
          },
          "canUseModerationTools": {
            "type": "boolean"
          },
          "fromLanguage": {
            "type": "string"
          },
          "_achievements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Achievement"
            }
          },
          "location": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "hasGoogleId": {
            "type": "boolean"
          },
          "optionalFeatures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OptionalFeature"
            }
          },
          "currentCourse": {
            "$ref": "#/components/schemas/CurrentCourse"
          },
          "currentCourseId": {
            "type": "string"
          },
          "hasPhoneNumber": {
            "type": "boolean"
          },
          "creationDate": {
            "type": "integer"
          },
          "hasPlus": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emailVerified": {
            "type": "boolean"
          },
          "courses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "xpGains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/XpGain"
            }
          },
          "enableSpeaker": {
            "type": "boolean"
          },
          "deactivated": {
            "type": "boolean"
          },
          "emailEditSuggested": {
            "type": "boolean"
          },
          "emailStreakFreezeUsed": {
            "type": "boolean"
          },
          "gems": {
            "type": "integer"
          },
          "pushEditSuggested": {
            "type": "boolean"
          },
          "pushAnnouncement": {
            "type": "boolean"
          },
          "totalXp": {
            "type": "integer"
          },
          "enableMicrophone": {
            "type": "boolean"
          },
          "insiteImmersionLingots": {
            "type": "boolean"
          },
          "pushPromotion": {
            "type": "boolean"
          },
          "emailStreakSaver": {
            "type": "boolean"
          },
          "insiteSentenceEdited": {
            "type": "boolean"
          },
          "canCreatePearsonClassroom": {
            "type": "boolean"
          },
          "xpGoal": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "timezoneOffset": {
            "type": "string"
          },
          "coachOutfit": {
            "type": "string"
          },
          "weeklyXp": {
            "type": "integer"
          },
          "lingots": {
            "type": "integer"
          },
          "googleId": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "streakData": {
            "$ref": "#/components/schemas/StreakData"
          },
          "referralInfo": {
            "$ref": "#/components/schemas/ReferralInfo"
          },
          "pushStreamPost": {
            "type": "boolean"
          },
          "adsEnabled": {
            "type": "boolean"
          },
          "emailAnnouncement": {
            "type": "boolean"
          },
          "xpGoalMetToday": {
            "type": "boolean"
          },
          "pushLeaderboards": {
            "type": "boolean"
          },
          "zapsRemaining": {
            "type": "integer"
          },
          "emailSchoolsAnnouncement": {
            "type": "boolean"
          },
          "emailWeeklyProgressReport": {
            "type": "boolean"
          },
          "emailWeeklyReport": {
            "type": "boolean"
          },
          "pushComment": {
            "type": "boolean"
          },
          "pushFollow": {
            "type": "boolean"
          },
          "pushClubs": {
            "type": "boolean"
          },
          "pushStreakSaver": {
            "type": "boolean"
          },
          "emailClassroomJoin": {
            "type": "boolean"
          },
          "emailAssignment": {
            "type": "boolean"
          },
          "emailAssignmentComplete": {
            "type": "boolean"
          },
          "lastStreak": {
            "$ref": "#/components/schemas/definitions-LastStreak"
          },
          "shopItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShopItem"
            }
          },
          "health": {
            "$ref": "#/components/schemas/Health"
          }
        }
      },
...
    }
  }
}